Refactored to flatten out Cart. Removed Selectquantity
